add show button Task component
