+++ services/app/assets/js/widgets/config/editorSettingsByUserType.js
@@ -3,7 +3,6 @@ import editorUserTypes from './editorUserTypes';
 const defaultSettings = {
     editable: false,
     showControlBtns: false,
-    showGuideBtnStatus: 'disabled',
     checkBtnStatus: 'disabled',
     resetBtnStatus: 'disabled',
     langPickerStatus: 'disabled',
@@ -15,7 +14,6 @@ export default {
   [editorUserTypes.currentUser]: {
     editable: true,
     showControlBtns: true,
-    showGuideBtnStatus: 'enabled',
     checkBtnStatus: 'enabled',
     resetBtnStatus: 'enabled',
     giveUpBtnStatus: 'enabled',
