+++ services/app/assets/js/widgets/containers/RootContainer.jsx
@@ -1,10 +1,9 @@
-import React, { useEffect, useContext, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect, useSelector, useDispatch } from 'react-redux';
 import Gon from 'gon';
 import ReactJoyride, { STATUS } from 'react-joyride';
 import { CSSTransition, SwitchTransition } from 'react-transition-group';
-import _ from 'lodash';
 import { useMachine } from '@xstate/react';
 
 import GameWidget from './GameWidget';
@@ -15,8 +14,6 @@ import { actions } from '../slices';
 import * as GameActions from '../middlewares/Game';
 import * as ChatActions from '../middlewares/Chat';
 import {
-  gamePlayersSelector,
-  currentUserIdSelector,
   isShowGuideSelector,
 } from '../selectors';
 import WaitingOpponentInfo from '../components/WaitingOpponentInfo';
@@ -112,18 +109,11 @@ const steps = [
 
 const GameWidgetGuide = () => {
   const dispatch = useDispatch();
-  const { current } = useContext(GameContext);
   const [isFirstTime, setIsFirstTime] = useState(window.localStorage.getItem('guideGamePassed') === null);
-  const isActiveGame = current.matches({ game: 'active' });
-  const players = useSelector(state => gamePlayersSelector(state));
-  const currentUser = useSelector(state => currentUserIdSelector(state));
   const isShowGuide = useSelector(state => isShowGuideSelector(state));
-  const isCurrentPlayer = _.has(players, currentUser);
 
   return (
-    (isShowGuide || isFirstTime)
-    && isActiveGame
-    && isCurrentPlayer && (
+    (isShowGuide || isFirstTime) && (
       <ReactJoyride
         continuous
         run
